+++ CHANGES.md
@@ -1,6 +1,8 @@
 Next Release
 ===========
 * Remove the standalone version of the emulator.
+* Add an option to enable/disable printer emulation.
+* Add a virtual keyboard to play without a keyboard.
 
 Release 0.8 (2018/04/23)
 ===========